+++ platform_flag.go
@@ -133,7 +133,9 @@ func (s *appendPlatformValue) Set(value string) error {
 	}
 
 	for _, v := range strings.Split(value, " ") {
-		s.appendIfMissing(strings.ToLower(v))
+		if v != "" {
+			s.appendIfMissing(strings.ToLower(v))
+		}
 	}
 
 	return nil