+++ platform_flag_test.go
@@ -131,6 +131,14 @@ func TestAppendPlatformValue_impl(t *testing.T) {
 func TestAppendPlatformValue(t *testing.T) {
 	var value appendPlatformValue
 
+	if err := value.Set(""); err != nil {
+		t.Fatalf("err: %s", err)
+	}
+
+	if len(value) > 0 {
+		t.Fatalf("bad: %#v", value)
+	}
+
 	if err := value.Set("windows LINUX"); err != nil {
 		t.Fatalf("err: %s", err)
 	}